Steady-state scheduling
What we are actually trying to determine is SDP efficiency over a long
time period, for which it might not be sufficient to just take the
average over a couple of weeks. This implements a compromise: We
basically repeat the same sequence over and over until we reach a
fixpoint, then take the average over that.