Scale A-kernel support depending on NIpatches and Nfacet
As the A-kernel needs to reflect the direction-dependence of
ionospheric calibration it must have at least as many pixels as
directions we solve calibration in. On the other hand, with facetting
we are imaging different parts of the sky separately, and therefore
only part of that complexity applies to individual A-kernels (we can
sample it more coarsely in grid space).

The new formula attempts to represent this fairly: We get a certain
base kernel size for anti-aliasing (constant, as it relates to the
facet size) plus a contribution from the number of i-patches
overlapping the facet. This is assumed to be roughly the total number
of patches divided by the number of facets. We multiply this by 2 to
account for some over-sampling and uneven distribution of i-patches.